Revert "feat: wip render for sizes calculates once per column"
This reverts commit fddf930ad87f751f47f6d19de23ba7d0f4529548.
