+++ src/Molecules/FlexTable/Cell/Cell.tsx
@@ -42,7 +42,6 @@ const Cell: CellComponent = (props) => {
       md={mdTable}
       lg={lgTable}
       xl={xlTable}
-      columnId={columnId}
     >
       {({ fontSize, className: mediaClassName }) => (
         <InnerCell
