+++ src/Molecules/FlexTable/Footer/Footer.tsx
@@ -32,7 +32,6 @@ const Footer: FooterComponent = (props) => {
       md={mdTable}
       lg={lgTable}
       xl={xlTable}
-      columnId={columnId}
     >
       {({ fontSize, className: mediaClassName }) => (
         <StyledFlexbox
